requirements.txt

opencv-python==4.9.0.80
Pillow==10.3.0
pandas==2.2.2
numpy==1.26.4
